chore: moved all sub reducers into a reducers directory under the store directory
