+++ client/store/reducers/layoutReducer.js
